+++ src/routes/mission.tsx
@@ -1,45 +1,9 @@
 import { Button, Card } from '@nextui-org/react';
 
 import LayoutContainer from 'components/layout/LayoutContainer';
-import Task, { type Props as TaskItem } from 'components/Task';
+import Task from 'components/Task';
 
-const tasks: TaskItem[] = [
-  {
-    title: 'Share Webtoon 5 Times or More',
-    point: 100,
-    currentMissionCount: 1,
-    maxMissionCount: 5,
-    status: 'ongoing',
-  },
-  {
-    title: 'Click Ads 10 Times or More',
-    point: 100,
-    currentMissionCount: 8,
-    maxMissionCount: 10,
-    status: 'ongoing',
-  },
-  {
-    title: 'View Webtoon 40 Times or More',
-    point: 100,
-    currentMissionCount: 40,
-    maxMissionCount: 40,
-    status: 'ongoing',
-  },
-  {
-    title: 'Like Webtoon 20 Times or More',
-    point: 100,
-    currentMissionCount: 20,
-    maxMissionCount: 20,
-    status: 'completed',
-  },
-  {
-    title: 'Watch Ads 20 Times or More',
-    point: 100,
-    currentMissionCount: 1,
-    maxMissionCount: 1,
-    status: 'completed',
-  },
-];
+import tasks from 'data/task';
 
 const Mission = () => (
   <LayoutContainer>
